requirements.txt
pandas
langchain
langchain_experimental
langchain_openai
lida  
python-dotenv
fastapi
uvicorn
pydantic